AP-61 | Restructure tests to test useListViewAction separately *Remove snapshot tests, no benefit of testing CustomPopup. It is tested separately
