+++ ui/package.json
@@ -45,6 +45,7 @@
     "@storybook/react": "^5.2.4",
     "@testing-library/jest-dom": "^4.1.0",
     "@testing-library/react": "^9.1.4",
+    "@testing-library/react-hooks": "^3.4.1",
     "babel-core": "^6.26.3",
     "babel-jest": "^24.9.0",
     "babel-loader": "^8.0.6",
@@ -70,6 +71,7 @@
     "react-app-polyfill": "^1.0.4",
     "react-dev-utils": "^9.0.4",
     "react-select-event": "^2.1.1",
+    "react-test-renderer": "^16.13.1",
     "resolve-url-loader": "^3.1.0",
     "sass-loader": "^8.0.0",
     "style-loader": "^1.0.0",
